requirement.txt
antlr4-python3-runtime==4.8
asgiref==3.5.0
certifi==2021.10.8
chardet==3.0.4
charset-normalizer==2.0.12
Django==3.2.5
django-filter==21.1
djangorestframework==3.12.4
hydra-sdk==1.0.2
idna==2.8
importlib-resources==5.6.0
omegaconf==2.1.1
python-decouple==3.6
pytz==2022.1
PyYAML==6.0
requests==2.22.0
sqlparse==0.4.2
urllib3==1.25.11
zipp==3.7.0
